requirements.txt
vk-api==11.8.0
python-telegram-bot==12.7
